Module Hardcaml_waveterm

Hardcaml_waveterm is a library for displaying terminal-based waveforms from hardcaml simulations.

module Display_rule : sig ... end

A Rule.t is a predicate on Port.ts that specifies the display format of matching ports.

module Display_rules : sig ... end

A Display_rules.t is an ordered list of rules that specifies the order of ports and the formatting of signals in a waveform.

module Text_alignment : sig ... end

module Wave_format : sig ... end

Display format for a single wave (a row in the complete waveform).

module Waveform : sig ... end

A waveform is a mutable value that represents the outcome of a simulation; it holds a time series of values for a number of ports. The waveform is updated by running the simulation.

module Expert : sig ... end