Module type Draw.S

Main graphics drawing API.

Most functions take a bounds parameter which is a rectangle to which drawing is clipped and also serves as an origin for any coordinates.

type ctx

drawing context

type style

underlying style type

val rows : ctx -> Base.int

val cols : ctx -> Base.int

val get_bounds : ctx -> rect

get context size

val get_style : Style.t -> style

convert our style info to underlying style

val clear : ctx -> Base.unit

clear display

val fill : ctx:ctx -> style:style -> bounds:rect -> Base.char -> Base.unit

fill bounds with char given style

val draw_int : ctx:ctx -> style:style -> bounds:rect -> r:Base.int -> c:Base.int -> Base.int -> Base.unit

draw int (representing unicode value)

val draw_piece : ctx:ctx -> style:style -> bounds:rect -> r:Base.int -> c:Base.int -> piece -> Base.unit

draw piece

val draw_char : ctx:ctx -> style:style -> bounds:rect -> r:Base.int -> c:Base.int -> Base.char -> Base.unit

draw char

val draw_string : ctx:ctx -> style:style -> bounds:rect -> r:Base.int -> c:Base.int -> Base.string -> Base.unit

draw string (nothing fancy - horizontal, no breaks)

val draw_box : ctx:ctx -> style:style -> bounds:rect -> Base.string -> Base.unit

draw box outline with label

val get : ctx:ctx -> bounds:rect -> r:Base.int -> c:Base.int -> Base.int * Style.t

get value and style at point

val inv : ctx:ctx -> bounds:rect -> r:Base.int -> c:Base.int -> Base.unit

invert fg and bg at point

val bold : ctx:ctx -> bounds:rect -> r:Base.int -> c:Base.int -> Base.unit

set bold on point