Module Hardcaml_waveterm.Wave_format

Display format for a single wave (a row in the complete waveform).

type t =

| Binary	(*
Binary.
*)

| Bit	(*
Like Binary, but shows a single bit as a line.
*)

| Bit_or of t	(*
If width=1 then Bits otherwise t
*)

| Hex	(*
Hexadecimal.
*)

| Unsigned_int	(*
Unsigned integer.
*)

| Int	(*
Signed integer.
*)

| Index of Base.string Base.list	(*
Use data as index into given list of strings.
*)

| Custom of Hardcaml.Bits.t -> Base.string	(*
User defined formatting.
*)

val sexp_of_t : t -> Sexplib0.Sexp.t

Custom constructors are compared for physical equality.

include Base.Equal.S with type t := t

val equal : t Base__Equal.equal